Update README.md, changed to-do-list again
OSCP and CEH currently have the most job postings listed, so they take priority.

HackTheBox has their own private career network and is known in the industry, so it closely follows.

The overwhelming consensus is that CEH is useless for gaining actual skills to pass a technical interview and is mostly an HR buffer, so it's something I should only look into getting when I'm job-ready in my mind. That's why it comes after OSCP. It opens up more job opportunities than the somewhat pitiful job openings that HackTheBox occasionally gives, though, so it's more important for getting a career sooner and comes before it.

OSCP is hard but the eLearnSecurity pathway has been pointed out by a few people as a great build up to it with a lot to learn, plus I already have a subscription to them since I'm working on eJPT. It's also the gold standard for both technical prowess and HR, having many job postings listing it as a requirement, so getting it should be my main focus. From what I read, it seems like the actual course requires the trainee to do research outside of it and already have a decent foundation in cybersecurity, so taking other courses that will fill in those information gaps will help a lot. Otherwise, I would just jump straight into OSCP.

Taking these courses seriously in this order should teach me what I need to know to start in the industry. I'm glad I went ahead to get Tech Support Fundamentals and study Security+, because they gave a great foundation.