Update README.md, added special interests and links
This time I didn't get too carried away enumerating every loosely affiliated exploit I could think of to get the point across. I always knew that I was interested in low-level hacking but I lacked the terminology to express that, so now I can phrase it in a way that is cleaner and more accurate.

I could probably simplify to just "operating system exploitation" or "host-based exploitation." The problem with the former is that not everything considered a part of the OS is low-level, and is therefore less interesting to me, and the problem with the latter is that I'm just as interested in exploiting network interface cards and router operating systems, too, which is arguably a part of network exploitation rather than host-based exploitation.

I also added links to many of the coding projects. I might come back and add links to the CTFs and certificate programs as well. This is mostly so I can continue using this README as a to-do list, but it does also provide context to what I'm talking about.

I might want to specify what I mean by "coding principles" in a future update or in an entirely separate repository. For now, remember that the idea is to demonstrate versatile computer science concepts between languages, such as proper input validation, basic sorting algorithms, etc.