Update README.md, added more projects
These projects are getting a little out of hand. I'm not sure they function as a straightforward to-do list anymore, because I have half a mind to skip down Linux From Scratch and go from there.

I've been so focused on eJPT and I've been so sick that I haven't had the time or mental resources to work on these other projects. I can still work from when I wake up to when I sleep, but my progress is a lot slower than it used to be. I just spent two weeks out of commission due to starting immunotherapy, which put me more behind schedule than I was hoping for.

I am frustrated by these constant setbacks and the growing list of what I want to accomplish. The upside is that I know exactly what I want to focus on, what I want to achieve, and how to go about doing it. This is a level of clarity and passion that I think most people can only dream of, and I'm basically certain that I won't ever experience a mid-life crisis or anything that will dramatically alter my life trajectory later on, either.

I'm trying to remember that the pandemic severely hindered most people my age. I'm not the only one that had my college life destroyed by those events, even though many people my age were still somehow able to continue mostly unaffected.

I was impacted to a greater degree, and it's not like I was on the right course before the pandemic, either, but I think I can pull all of this off. It only feels overwhelming when I look at this list as a whole instead of focusing on each element step-by-step, because this list will probably take at least a couple of years to finish, if not more.

I just have to pace myself and remember that it's a marathon, not a race. I'll be able to accomplish larger workloads as I get back into the swing of things, because hopefully the days of chronic illness are finally behind me now.

I've been reading up on how people in East Asian countries cope with higher workloads and are able to work essentially from waking to sleeping, and I think I can do that, too. It's going to be a little exhausting at first, but as long as I pace myself and focus on just doing what I can to stay on course, I should adapt to the workload.

I don't know how much of an advantage that will give me in the offensive security scene. I know it's a hyper-competitive field as it is. I'm trying to not worry about it and just focus on the work in front of me. Everywhere I turn online, people are highly pessimistic and negative about getting a pentesting job with no experience, but the actual people in the industry that I've connected to have told me that I can do it, which is more encouraging.

It sometimes feels like a futile struggle, still, but I'm not going to give up. This is what I want from life. I would never forgive myself if I didn't put 110% into achieving it. So I'll add more projects, knowing that I WILL get to them some day, even if they feel out of reach right now. I will look at this list every few days to remember what I'm striving towards.

You hear that, DEFCON and Jack Rhysider? I'm a nobody now, but one day it will be impossible to ignore what I've accomplished. It's only a matter of time.

...I hope.