Update README.md, added quantum programming to list
Quantum computing is almost inevitable at this rate, so I should prepare for it.

I honestly prefer the machine code of traditional binary, because it has Boolean logic baked into it. It's more black-and-white than quantum programming is.

Nonetheless, there are low-level quantum programming languages that are comparable to Assembly, and they will likely become more common over the next couple of decades. They might not enter mainstream use until 2040, and civilization might collapse by 2050, so they're not as important to be familiarized with. Despite that, it would be smart to prepare when I have the time, while I'm learning traditional Assembly, anyway.

The existence of quantum computers will make low-level hacking more complex than it already is, which is honestly pretty scary to me, and just the idea of a "quantum" computer already sounds esoteric enough to avoid dark alleys that one might be lurking in. I won't get anywhere by avoiding the topic in the hope that it will go away, though.

Hopefully, by the time I get into quantum exploitation, I will already be experienced enough in low-level hacking that it will feel more like another tool in my belt than a complete reinvention of the wheel from scratch. I might end up being too late on that bandwagon, though, because I still have to catch up on traditional low-level hacking.

Agh! Tech moves so fast. It's hard to stay oriented. I hate it. I also love it, because it means there's always more to learn, but I hate it because it means I have to constantly adapt to stay relevant in the job market. I have to foresee where the industry will be in half a decade because it will take me half a decade to learn the skills I need to enter it, and by the time I gain those skills they might no longer be relevant anymore.

It feels like a huge game of chance. And that's on top of the fact that offensive cybersecurity is an extremely competitive field to begin with. You know, I don't like risk, I hate conflict, and I'm not a masochist. Why am I doing this to myself??

Well, I'll tell you why. I like hacking. If I have to learn low-level quantum programming, prove myself to HR, defy the gatekeeping elitists, and beat out my competitors then that's what I have to do to keep hacking. I don't like it. I'd rather avoid that hassle entirely. But to keep doing what I love, the sacrifice is worth it.

It hurts my feelings to have to fight everyone. It's stressful and worrying to have to constantly keep up with a field that's rapidly developing in multiple directions. It makes me feel worthless and inadequate to see so many people who know way more than me and are so much further ahead than I am.

But I'm not going to give up. That would be pathetic. I'm in this for the long haul, for better or worse.