Update README.md, removed "malicious packet injection"
First off, "packet injection" by itself would have worked.

Secondly, I'm really more interest in remote kernel exploitation than packet injection, which is a separate concept.

From what I can find online, there are examples of packet injection being used for code injection, but it's mostly used for on-path attacks and DoS attacks, which I'm less interested in. Not uninterested in! They aren't beneath me! Just less interested.

Thirdly, it's nice to simplify back to just kernel exploitation. No chaining from packet injection necessary! Or, in a sense, packet injection is a part of kernel exploitation, sort of.

Dan Rosenberg used what appears to be an injected packet (an evil ROSE frame compiled and sent using rose_call) in his remote kernel exploit to cause an overflow and gain control of the EIP. This is the part of the kernel exploit that makes it doable remotely and it's done by abusing a network protocol.

From what I can tell, packet injection is one common piece of a broader remote kernel exploit. So I don't need to specify packet injection + kernel exploitation. That's covered under "remote kernel exploitation."

Fourthly, this simplification to a single phrase is nice because it makes it feel like it's less intimidating to me. Consolidating the concept into one idea is useful. In my mind, it means I can hyperfocus on "remote kernel exploits" rather than "kernel exploits" and "packet injection" even if "remote kernel exploits" include forms of packet injection.

Fifthly, it's also more communicative, because I'm not interested in packet injection as a general concept. My interest in it is mostly as a means of kernel exploitation. As I said, I'm not especially interested in DoS attacks or MITM attacks or even cheating in mobile phone games.

It's not that packet injection isn't interesting to me. It is, and I even performed a (simulated) MITM attack today in the INE's PTS. Obviously, I have to understand it and put it in my utility belt to be a pentester or a red teamer, and I'll give it the proper respect it's due.

It's just that this part of my bio is meant to showcase what I'm most interested in, and what really fascinates me the most is kernel exploitation. I have nothing against packet injection. I'm just more interested in kernel exploits.

So this might be me inadvertently showing the depths of my ignorance, but as I make this commit I still consider myself a "green hat" and a noob, so that's fine. If this update does turn out to be particularly dumb on my part, future updates will hopefully show how much I've grown.

...of course, nobody reads these except me, so this concern is a little neurotic.