Update README.md, kernels, shells, and binaries
I think this is probably the best way that I can word it.

I keep rewording this entry and it might never be totally satisfying to me, but I keep thinking of a better way to describe what I'm trying to say in this opening paragraph.

I would just say "low-level bug bounty hunting," but I think that's less precise and I think it can give the wrong idea. "Low-level" can be mistaken for "easy" or "elementary," which is almost the opposite of what it actually means. "Bug bounty hunting" also puts a bit less emphasis on actual exploitation.