Update README.md, removed Security+
That exam isn't happening. I studied the material well-enough to pass questions about it in an interview, but a lot of job listings aren't even asking for it. I did see a few that listed PenTest+, which was weird because that was a completely useless cert a few years ago.

Hopefully I can catch up to the industry as my life stabilizes. We'll see.