Update README.md, revised coding-principles
LeetCode challenges can be used as a demonstration of how well I know various programming languages. Which, as of this moment, is absolutely not at all.

C and Python seem to be enduring languages, both show up in a lot of places. Rust is a major one when it comes to memory, and I'm still interested in memory hacking so I'm keeping it in mind.

I don't know how many LeetCode challenges to do or which ones to prioritize. I guess maybe the first 23, so the first three easy-to-hard segments? Start there and move on, maybe.

Start with C. Then Python. Then Rust. That seems systematic enough. Then I can wrap back around after I'm done with that.