Update README.md, no longer interested in CTFs
I've been disillusioned with CTFs a little bit. Many of them cost money to participate in, and their applicability to real-world scenarios is widely debated and variable. They tend to not be taken seriously as experience, except through HackTheBox job offerings, which you need money to participate in and you need to work up to.

If they aren't a scam, they're for people with far more economic privilege than me.

I'm a lot more interested in developing my skills with vulnerability research, bug bounty hunting, and exploit development. These programs will pay me if I'm good at them, and there are a lot of free resources on them.

These are actually not that much more complicated than being a total script kiddie, in theory. The only difference is that you're searching for code vulnerabilities rather than configuration vulnerabilities. It's the same general methodology of finding and exploiting vulnerabilities you see in junior pentesting or vulnerability auditing, simply more advanced.