Update README.md
Vulnerability research is considered to be a part of exploit development. It's also often done using low-level programming with systems and network services, so my older mention of systems and networking and low-level programming are implied by this, more or less.

I'm trying to streamline my profile here. That's all.